Added auth to UserComicEndpoints Fixed validation for user trying to delete

+++ BooksAPI/BooksAPI.BE/Endpoints/UserComicEndpoints.cs
@@ -1,4 +1,6 @@
-﻿using BooksAPI.BE.Contracts.UserComic;
+﻿using System.Security.Claims;
+using BooksAPI.BE.Constants;
+using BooksAPI.BE.Contracts.UserComic;
 using BooksAPI.BE.Entities;
 using BooksAPI.BE.Exception;
 using BooksAPI.BE.Interfaces.Repositories;
@@ -15,15 +17,17 @@ public static class UserComicEndpoints
 {
     public static void MapUserComicEndpoints(this WebApplication app)
     {
-        app.MapPost("/userComic", CreateUserComic);
+        app.MapPost("/userComic", CreateUserComic).RequireAuthorization(AppConstants.PolicyNames.UserRolePolicyName);
 
-        app.MapGet("/userComic/{id:guid}", GetUserComicById);
-        app.MapGet("/userComic/", GetAllUserComics);
-        app.MapGet("/userComics/", GetAllUserComicsByUserId);
+        app.MapGet("/userComic/{id:guid}", GetUserComicById)
+            .RequireAuthorization(AppConstants.PolicyNames.UserRolePolicyName);
+        app.MapGet("/userComic/", GetAllUserComics).RequireAuthorization(AppConstants.PolicyNames.AdminRolePolicyName);
+        app.MapGet("/userComics/", GetAllUserComicsByUserId)
+            .RequireAuthorization(AppConstants.PolicyNames.UserRolePolicyName);
 
-        app.MapPut("/userComic/", UpdateUserComic);
+        app.MapPut("/userComic/", UpdateUserComic).RequireAuthorization(AppConstants.PolicyNames.UserRolePolicyName);
 
-        app.MapDelete("/userComic/", DeleteUserComic);
+        app.MapDelete("/userComic/", DeleteUserComic).RequireAuthorization(AppConstants.PolicyNames.UserRolePolicyName);
     }
 
     public static void AddUserComicServices(this IServiceCollection services)
@@ -31,9 +35,10 @@ public static class UserComicEndpoints
         services.AddScoped<IUserComicRepository, UserComicRepository>();
         services.AddScoped<IUserComicService, UserComicService>();
         services.AddScoped<IValidator<UserComic>, UserComicValidator>();
+        services.AddHttpContextAccessor();
     }
 
-    static async Task<IResult> CreateUserComic(IUserComicService service, [FromBody]CreateUserComicRequest request)
+    static async Task<IResult> CreateUserComic(IUserComicService service, [FromBody] CreateUserComicRequest request)
     {
         try
         {
@@ -58,7 +63,7 @@ public static class UserComicEndpoints
         }
     }
 
-    static async Task<IResult> GetUserComicById(IUserComicService service, [FromRoute]Guid id)
+    static async Task<IResult> GetUserComicById(IUserComicService service, [FromRoute] Guid id)
     {
         try
         {
@@ -88,7 +93,7 @@ public static class UserComicEndpoints
         }
     }
 
-    static async Task<IResult> GetAllUserComicsByUserId(IUserComicService service, [FromQuery]string userId)
+    static async Task<IResult> GetAllUserComicsByUserId(IUserComicService service, [FromQuery] string userId)
     {
         try
         {
@@ -105,7 +110,8 @@ public static class UserComicEndpoints
         }
     }
 
-    static async Task<IResult> UpdateUserComic(IUserComicService service, [FromQuery]Guid id, [FromBody]UpdateUserComicRequest request)
+    static async Task<IResult> UpdateUserComic(IUserComicService service, [FromQuery] Guid id,
+        [FromBody] UpdateUserComicRequest request)
     {
         try
         {
@@ -129,13 +135,30 @@ public static class UserComicEndpoints
             return Results.StatusCode(StatusCodes.Status500InternalServerError);
         }
     }
-    
-    static async Task<IResult> DeleteUserComic(IUserComicService service, [FromQuery]Guid id, [FromQuery]string userId)
+
+    static async Task<IResult> DeleteUserComic([FromQuery] Guid id, [FromQuery] string userId,
+        IUserComicService service, HttpContext httpContext)
     {
         try
         {
+            var userIdFromAuth = GetUserIdFromAuth(httpContext);
+
+            if (userIdFromAuth == null)
+            {
+                return Results.Unauthorized();
+            }
+
+            if (userId != userIdFromAuth)
+            {
+                return Results.Forbid();
+            }
+
             await service.DeleteUserComic(id, userId);
             return Results.Ok();
+        }
+        catch (InvalidOperationException ex)
+        {
+            return Results.BadRequest(ex.Message);
         }
         catch (UserNotFoundException ex)
         {
@@ -150,5 +173,19 @@ public static class UserComicEndpoints
             return Results.StatusCode(StatusCodes.Status500InternalServerError);
         }
     }
-    
+
+    private static string? GetUserIdFromAuth(HttpContext httpContext)
+    {
+        ClaimsPrincipal user = httpContext.User;
+        string? userIdFromAuth = null;
+        foreach (Claim userClaim in user.Claims)
+        {
+            if (userClaim.Type == ClaimTypes.NameIdentifier)
+            {
+                userIdFromAuth = userClaim.Value;
+            }
+        }
+
+        return userIdFromAuth;
+    }
 }
